+++ nds-client/src/test/java/br/com/abril/nds/fixture/Fixture.java
@@ -1,6 +1,7 @@
 package br.com.abril.nds.fixture;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
@@ -36,7 +37,6 @@ import br.com.abril.nds.model.cadastro.GrupoFornecedor;
 import br.com.abril.nds.model.cadastro.GrupoProduto;
 import br.com.abril.nds.model.cadastro.LicencaMunicipal;
 import br.com.abril.nds.model.cadastro.MaterialPromocional;
-import br.com.abril.nds.model.cadastro.Moeda;
 import br.com.abril.nds.model.cadastro.OperacaoDistribuidor;
 import br.com.abril.nds.model.cadastro.ParametroCobrancaCota;
 import br.com.abril.nds.model.cadastro.ParametroContratoCota;
@@ -488,7 +488,7 @@ public class Fixture {
 	
 	public static Lancamento lancamento(TipoLancamento tipoLancamento,
 			ProdutoEdicao produtoEdicao, Date dlp, Date drp, Date dataCriacao,
-			Date dataStatus, BigDecimal reparte, StatusLancamento statusLancamento,
+			Date dataStatus, BigInteger reparte, StatusLancamento statusLancamento,
 			ItemRecebimentoFisico recebimento, Integer sequenciaMatriz) {
 		Lancamento lancamento = new Lancamento();
 		lancamento.setDataCriacao(dataCriacao);
@@ -530,7 +530,7 @@ public class Fixture {
 			StatusAprovacao statusAprovacao,
 			Date drd, 
 			Date dataStatus, 
-			BigDecimal qtde) {
+			BigInteger qtde) {
 		
 		ConferenciaEncalheParcial conferenciaEncalheParcial = new ConferenciaEncalheParcial();
 		
@@ -597,7 +597,7 @@ public class Fixture {
 	
 	public static Lancamento lancamentoExpedidos(TipoLancamento tipoLancamento,
 			ProdutoEdicao produtoEdicao, Date dlp, Date drp, Date dataCriacao,
-			Date dataStatus, BigDecimal reparte, StatusLancamento statusLancamento,
+			Date dataStatus, BigInteger reparte, StatusLancamento statusLancamento,
 			ItemRecebimentoFisico recebimento,Expedicao expedicao, Integer sequenciaMatriz) {
 		Lancamento lancamento = new Lancamento();
 		lancamento.setDataCriacao(dataCriacao);
@@ -712,7 +712,7 @@ public class Fixture {
 		return cota;
 	}
 
-	public static Estudo estudo(BigDecimal qtdReparte, Date data,
+	public static Estudo estudo(BigInteger qtdReparte, Date data,
 			ProdutoEdicao produtoEdicao) {
 
 		Estudo estudo = new Estudo();
@@ -726,8 +726,8 @@ public class Fixture {
 		return estudo;
 	}
 
-	public static EstudoCota estudoCota(BigDecimal qtdePrevista,
-			BigDecimal qtdeEfetiva, Estudo estudo, Cota cota) {
+	public static EstudoCota estudoCota(BigInteger qtdePrevista,
+			BigInteger qtdeEfetiva, Estudo estudo, Cota cota) {
 
 		EstudoCota estudoCota = new EstudoCota();
 
@@ -1046,7 +1046,7 @@ public class Fixture {
 	}
 	
 	public static ItemNotaFiscalEntrada itemNotaFiscal(ProdutoEdicao produtoEdicao,
-			Usuario usuario, NotaFiscalEntrada notaFiscal, Date dataLancamento, Date dataRecolhimento, TipoLancamento tipoLancamento, BigDecimal qtde) {
+			Usuario usuario, NotaFiscalEntrada notaFiscal, Date dataLancamento, Date dataRecolhimento, TipoLancamento tipoLancamento, BigInteger qtde) {
 		ItemNotaFiscalEntrada itemNotaFiscal = new ItemNotaFiscalEntrada();
 		itemNotaFiscal.setOrigem(Origem.MANUAL);
 		itemNotaFiscal.setProdutoEdicao(produtoEdicao);
@@ -1066,7 +1066,7 @@ public class Fixture {
 			Date dataLancamento, 
 			Date dataRecolhimento, 
 			TipoLancamento tipoLancamento, 
-			BigDecimal qtde,
+			BigInteger qtde,
 			String 		NCMProduto,
 			String 		CFOPProduto,
 			Long 		unidadeProduto,
@@ -1108,7 +1108,7 @@ public class Fixture {
 	public static ItemNotaFiscalSaida itemNotaFiscalSaida(
 			ProdutoEdicao produtoEdicao,
 			NotaFiscalSaida notaFiscal, 
-			BigDecimal qtde) {
+			BigInteger qtde) {
 		
 		ItemNotaFiscalSaida itemNotaFiscal = new ItemNotaFiscalSaida();
 		
@@ -1123,7 +1123,7 @@ public class Fixture {
 	public static ItemNotaFiscalSaida itemNotaFiscalSaidaNFE(
 			ProdutoEdicao produtoEdicao,
 			NotaFiscalSaida notaFiscal, 
-			BigDecimal qtde,
+			BigInteger qtde,
 			String		NCMProduto,
 			String 		CFOPProduto,
 			Long 		unidadeProduto,
@@ -1673,7 +1673,7 @@ public class Fixture {
 
 	public static ItemRecebimentoFisico itemRecebimentoFisico(ItemNotaFiscalEntrada itemNotaFiscal, 
 															  RecebimentoFisico recebimentoFisico,
-															  BigDecimal qtdeFisico) {
+															  BigInteger qtdeFisico) {
 		ItemRecebimentoFisico itemRecebimentoFisico = new ItemRecebimentoFisico();
 		itemRecebimentoFisico.setItemNotaFiscal(itemNotaFiscal);
 		itemRecebimentoFisico.setQtdeFisico(qtdeFisico);
@@ -1681,7 +1681,7 @@ public class Fixture {
 		return itemRecebimentoFisico;
 	}
 	
-	public static EstoqueProduto estoqueProduto(ProdutoEdicao produtoEdicao, BigDecimal qtde) {
+	public static EstoqueProduto estoqueProduto(ProdutoEdicao produtoEdicao, BigInteger qtde) {
 		EstoqueProduto estoqueProduto = new EstoqueProduto();
 		estoqueProduto.setProdutoEdicao(produtoEdicao);
 		estoqueProduto.setQtde(qtde);
@@ -1690,7 +1690,7 @@ public class Fixture {
 		return estoqueProduto;
 	}
 	
-	public static EstoqueProdutoCota estoqueProdutoCota(ProdutoEdicao produtoEdicao, BigDecimal qtde,
+	public static EstoqueProdutoCota estoqueProdutoCota(ProdutoEdicao produtoEdicao, BigInteger qtde,
 			Cota cota, List<MovimentoEstoqueCota> movimentos) {
 		EstoqueProdutoCota estoqueProdutoCota = new EstoqueProdutoCota();
 		estoqueProdutoCota.setCota(cota);
@@ -1709,7 +1709,7 @@ public class Fixture {
 													Usuario usuario, 
 													EstoqueProduto estoqueProduto,
 													Date dataInclusao,
-													BigDecimal qtde, 
+													BigInteger qtde, 
 													StatusAprovacao status,
 													String motivo) {
 
@@ -1742,7 +1742,7 @@ public class Fixture {
 		return ps;
 	}
 	
-	public static Diferenca diferenca(BigDecimal qtde,
+	public static Diferenca diferenca(BigInteger qtde,
 									  Usuario usuarioResponsavel,
 									  ProdutoEdicao produtoEdicao,
 									  TipoDiferenca tipoDiferenca,
@@ -1766,8 +1766,8 @@ public class Fixture {
 	}
 	
 	public static EstoqueProdutoCota estoqueProdutoCota(
-			ProdutoEdicao produtoEdicao, Cota cota, BigDecimal qtdeRecebida,
-			BigDecimal qtdeDevolvida) {
+			ProdutoEdicao produtoEdicao, Cota cota, BigInteger qtdeRecebida,
+			BigInteger qtdeDevolvida) {
 		EstoqueProdutoCota estoqueProdutoCota = new EstoqueProdutoCota();
 		estoqueProdutoCota.setCota(cota);
 		estoqueProdutoCota.setProdutoEdicao(produtoEdicao);
@@ -1779,7 +1779,7 @@ public class Fixture {
 	public static MovimentoEstoqueCota movimentoEstoqueCota(
 			ProdutoEdicao produtoEdicao, TipoMovimentoEstoque tipoMovimento,
 			Usuario usuario, EstoqueProdutoCota estoqueProdutoCota,
-			BigDecimal qtde, Cota cota, StatusAprovacao statusAprovacao, String motivo) {
+			BigInteger qtde, Cota cota, StatusAprovacao statusAprovacao, String motivo) {
 
 		MovimentoEstoqueCota movimentoEstoque = new MovimentoEstoqueCota();
 		movimentoEstoque.setData(new Date());
@@ -1800,7 +1800,7 @@ public class Fixture {
 			TipoMovimentoEstoque tipoMovimento,
 			Usuario usuario, 
 			EstoqueProdutoCota estoqueProdutoCota,
-			BigDecimal qtde, 
+			BigInteger qtde, 
 			Cota cota, 
 			StatusAprovacao statusAprovacao, 
 			String motivo) {
@@ -1822,7 +1822,7 @@ public class Fixture {
 			MovimentoEstoqueCota movimentoEstoqueCota,
 			ChamadaEncalheCota chamadaEncalheCota,
 			ControleConferenciaEncalheCota controleConferenciaEncalheCota,
-			Date data, BigDecimal qtdeInformada, BigDecimal qtde,
+			Date data, BigInteger qtdeInformada, BigInteger qtde,
 			ProdutoEdicao produtoEdicao) {
 		
 		ConferenciaEncalhe conferenciaEncalhe = new ConferenciaEncalhe();
@@ -1843,7 +1843,7 @@ public class Fixture {
 			ChamadaEncalhe chamadaEncalhe,
 			boolean conferido,
 			Cota cota,
-			BigDecimal qtdePrevista) {
+			BigInteger qtdePrevista) {
 		
 		ChamadaEncalheCota chamadaEncalheCota = new ChamadaEncalheCota();
 		
@@ -1856,7 +1856,7 @@ public class Fixture {
 		
 	}
 	
-	public static RateioDiferenca rateioDiferenca(BigDecimal qtde, Cota cota, Diferenca diferenca, EstudoCota estudoCota){
+	public static RateioDiferenca rateioDiferenca(BigInteger qtde, Cota cota, Diferenca diferenca, EstudoCota estudoCota){
 		RateioDiferenca rateioDiferenca = new RateioDiferenca();
 		rateioDiferenca.setCota(cota);
 		rateioDiferenca.setDiferenca(diferenca);
@@ -1877,7 +1877,7 @@ public class Fixture {
 	
 	public static Lancamento lancamentos(TipoLancamento tipoLancamento,
 			ProdutoEdicao produtoEdicao, Date dlp, Date drp, Date dataCriacao,
-			Date dataStatus, BigDecimal reparte, StatusLancamento statusLancamento,
+			Date dataStatus, BigInteger reparte, StatusLancamento statusLancamento,
 			List<ItemRecebimentoFisico> recebimentos, Integer sequenciaMatriz) {
 		Lancamento lancamento = new Lancamento();
 		lancamento.setDataCriacao(dataCriacao);
@@ -2987,7 +2987,7 @@ public class Fixture {
 			Long ncm,
 			NotaFiscal notaFiscal,
 			ProdutoEdicao produtoEdicao,
-			BigDecimal quantidade,
+			BigInteger quantidade,
 			String unidade,
 			BigDecimal valorDesconto,
 			BigDecimal valorFrete,
